backend/requirements.txt
fastapi==0.112.0
uvicorn==0.21.1
motor==3.1.2
python-dotenv==1.0.0
google-genai==1.19.0
pillow==10.0.0
python-multipart==0.0.6
requests==2.31.0
